requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
boto3
s3fs
mlflow
python-dotenv
jupyter
